refactor: LottoStatistics 도메인 클래스 추가
